Library support
Define new target "libraries" that generates shared library for mfista
functions (dft and fft). Compile flag for libraries is bit different
from the one for executables. So, I defined suffix rule specific for
libraries (.c.o2). The target "libraries" generates two libraries:
libmfista_dft.so and libmfista_fft.so. The target is not included in
"all" so far. Also, libraries are not part of install target.